move notifications.js
